Revert isDesktopVideoPlayerSupported removal
