+++ packages/stream_chat_flutter/lib/src/utils/device_segmentation.dart
@@ -9,6 +9,11 @@ bool get isMobileDevice => CurrentPlatform.isIos || CurrentPlatform.isAndroid;
 /// Returns true if the app is running in a desktop device.
 bool get isDesktopDevice => CurrentPlatform.isMacOS || CurrentPlatform.isWindows || CurrentPlatform.isLinux;
 
+/// Returns true if the app is running on windows or linux platform.
+bool get isDesktopVideoPlayerSupported =>
+    // Dart VLC is not supported on MacOS.
+    !CurrentPlatform.isMacOS && (CurrentPlatform.isWindows || CurrentPlatform.isLinux);
+
 /// Returns true if the app is running in a mobile or web.
 bool get isMobileDeviceOrWeb => isWeb || isMobileDevice;
